Factor out LazySequences; add alias fields to Models.
In particular, LazySequences (the big one); this ended up needing to
restructure the fields mechanism completely. Also, a slightly unintended
side-effect in __init__.py is suddenly having to rearrange loads of
models so their names are available at the right time (or not, see the
LaterClass workaround, which I don't think is too bad).